Monero fee improvement attempt
The existing code creates a Monero fee in the UI of 1.9E-09 XMR. This is substantially less than a normal Monero network fee.

The `fee` response from `get_fee_estimate` is the "Amount of fees estimated per byte in atomic-units".

https://docs.getmonero.org/rpc-library/monerod-rpc/?h=get_fee_estimate#get_fee_estimate

I thus change `feeRatePerKb` to `feeAtomicRatePerByte`, and I use `feeRatePerByte` to convert the fee in atomic units to non-atomic. Then I multiply this value by 1500 bytes, which is currently a reasonable approximate weight of a 1-in/2-out Monero transaction.